// src/utils/types.ts
export interface AllCHaracters {
    count:    number;
    next:     string;
    previous: null;
    results:  Character[];
}

export interface Character {
    name:       string;
    height:     string;
    mass:       string;
    hair_color: string;
    skin_color: string;
    eye_color:  string;
    birth_year: string;
    gender:     Gender;
    homeworld:  string;
    films:      string[];
    species:    string[];
    vehicles:   string[];
    starships:  string[];
    created:    Date;
    edited:     Date;
    url:        string;
}
export interface CharacterWithImage {
    id:         string;
    name:       string;
    height:     string;
    mass:       string;
    hair_color: string;
    skin_color: string;
    eye_color:  string;
    birth_year: string;
    gender:     Gender;
    homeworld:  string;
    films:      string[];
    species:    string[];
    vehicles:   string[];
    starships:  string[];
    created:    Date;
    edited:     Date;
    url:        string;
    image:      string;
}

export enum Gender {
    Female = "female",
    Male = "male",
    NA = "n/a",
}